remove global.json from the template
I am unsure if I should remove this or not.  It signifies the version it was built with but versions above it should be fine.  I am removing it for now as I don't want users to have to have a specific version.  If the version they have is too old I suspect it will error out do to other dependencies. (Not sure)