Each of the Requests on Client we configured as if they returned the state they manipulated.
This was only used so I could determin which state to clone.  With this change all Action are nested classes of the state they change.  This gives us not only better encapsulation.
Each State effectively exposes its own API via Requests.  But it also give us the ability to determine the state be using decendentclass and thus not having to express it.  This simplifies the use case for developers.
The changes in this commit cascade from this reasoing.  We also introduce IAction as a marker class to be used for filters in middleware.  And rename RequestHandler to ActionHandler.
I think the overall result is cleaner and more consistent.