Breaking Change: Change Options to have Redux Dev Tools off by default.
Add conditional compile configurations for `ReduxDevToolsEnabled`
This includes a conditional compile flag.  We want redux dev tools as opt-in so I'm changing the default to false.
Add the Configuration to all the projects.
Bump Major version to 2 given breaking change set minor to 0.
Every Project gets this configuration, alghough I'm not sure how that should work.