The way packages seem to work now is different.
I don't have to embedd the static assets in a dll.  Just include them in the Nuget package staticwebassets folder which is done by putting them in the wwwroot folder in the project.  So now building the TypeScript project will generate the blazorstate.js file in the BlazorState projects wwwroot folder.

Accessing the js is now done using the Package-Id

Bump all the versions to the 3.1.0 release

In Startup we now need to UseStaticFiles prior to calling UseClienSideBlazorFiles.